# apilayer/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    MessageResponse:
        properties:
            Message:
                type: string
        title: MessageResponse ...
        type: object
        x-go-package: github.com/mohit2530/communityCare
...
    User:
        properties:
            InstanceID:
                format: uuid
                type: string
            aud:
                type: string
                x-go-name: Aud
            banned_until:
                format: date-time
                type: string
                x-go-name: BannedUntil
            confirmation_sent_at:
                format: date-time
                type: string
                x-go-name: ConfirmationSentAt
            confirmation_token:
                type: string
                x-go-name: ConfirmationToken
            confirmed_at:
                format: date-time
                type: string
                x-go-name: ConfirmedAt
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
            deleted_at:
                format: date-time
                type: string
                x-go-name: DeletedAt
            email:
                type: string
                x-go-name: Email
            email_change:
                type: string
                x-go-name: EmailChange
            email_change_confirm_status:
                format: int64
                type: integer
                x-go-name: EmailChangeConfirmStatus
            email_change_sent_at:
                format: date-time
                type: string
                x-go-name: EmailChangeSentAt
            email_change_token_current:
                type: string
                x-go-name: EmailChangeTokenCurrent
            email_change_token_new:
                type: string
                x-go-name: EmailChangeTokenNew
            email_confirmed_at:
                format: date-time
                type: string
                x-go-name: EmailConfirmedAt
            encrypted_password:
                type: string
                x-go-name: EncryptedPassword
            first_name:
                type: string
                x-go-name: FirstName
            id:
                format: uuid
                type: string
                x-go-name: ID
            invited_at:
                format: date-time
                type: string
                x-go-name: InvitedAt
            is_sso_user:
                type: boolean
                x-go-name: IsSSOUser
            last_name:
                type: string
                x-go-name: LastName
            last_sign_in_at:
                format: date-time
                type: string
                x-go-name: LastSignInAt
            middle_name:
                type: string
                x-go-name: MiddleName
            phone:
                type: string
                x-go-name: Phone
            phone_change:
                type: string
                x-go-name: PhoneChange
            phone_change_sent_at:
                format: date-time
                type: string
                x-go-name: PhoneChangeSentAt
            phone_change_token:
                type: string
                x-go-name: PhoneChangeToken
            phone_confirmed_at:
                format: date-time
                type: string
                x-go-name: PhoneConfirmedAt
            raw_app_meta_data:
                type: object
                x-go-name: RawAppMetaData
            raw_user_meta_data:
                type: object
                x-go-name: RawUserMetaData
            re_authentication_sent_at:
                format: date-time
                type: string
                x-go-name: ReAuthenticationSentAt
            re_authentication_token:
                type: string
                x-go-name: ReAuthenticationToken
            recovery_sent_at:
                format: date-time
                type: string
                x-go-name: RecoverySentAt
            recovery_token:
                type: string
                x-go-name: RecoveryToken
            role:
                type: string
                x-go-name: Role
            updated_at:
                format: date-time
                type: string
                x-go-name: UpdatedAt
        title: User ...
        type: object
        x-go-package: github.com/mohit2530/communityCare/model
...
info:
    description: |-
        # Documentation for the Asset alert api layer.

        Scehmes: http
    title: alert Api Layer
    version: 1.0.0
paths:
...
    /api/v1/logout:
        post:
            operationId: logout
            responses:
                "200":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Logs users out of the database system.'
            tags:
                - Logout
...
    /api/v1/profile/recent-activities:
        get:
            operationId: getRecentActivities
            parameters:
                - description: The userID of the selected user
                  in: query
                  name: id
                  required: true
                  type: string
                - description: The limit of recent activities.
                  format: int32
                  in: query
                  name: limit
                  required: true
                  type: integer
                - description: The timestamp with time zone until the data should be retrieved for
                  format: date-time
                  in: query
                  name: until
                  type: string
            responses:
                "200":
                    $ref: '#/responses/RecentActivitiesResponse'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Retrieves user actions for categories, maintenance plans, and assets. List all created, updated, or deleted.'
            tags:
                - GetRecentActivities
    /api/v1/signin:
        post:
            operationId: signin
            parameters:
                - description: The email address of the current user
                  in: query
                  name: email
                  required: true
                  type: string
                - description: The password of the current user
                  in: query
                  name: password
                  required: true
                  type: string
            responses:
                "200":
                    description: User
                    schema:
                        $ref: '#/definitions/User'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Sign in users into the database system.'
            tags:
                - Signin
    /api/v1/signup:
        post:
            operationId: signup
            parameters:
                - description: The email address of the current user
                  in: query
                  name: email
                  required: true
                  type: string
                - description: The password of the current user
                  in: query
                  name: password
                  required: true
                  type: string
                - description: The birthdate of the current user. Must be 13 years of age.
                  in: query
                  name: birthday
                  required: true
                  type: string
                - default: "false"
                  description: The user role for the application.
                  in: query
                  name: role
                  type: string
            responses:
                "200":
                    description: User
                    schema:
                        $ref: '#/definitions/User'
                "400":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "404":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
                "500":
                    description: MessageResponse
                    schema:
                        $ref: '#/definitions/MessageResponse'
            summary: '# Sign up users into the database system.'
            tags:
                - Signup
...
produces:
    - application/json
swagger: "2.0"